Eliminate StakeClaimNFTsView
